clisqlexec: drop the dependency on `tree`
This removes the remaining dependency on `tree` by using the function
from `lexbase` to escape strings, instead of `tree.DString`.

(The function first needs to move from `lex` to `lexbase`.)

Release note: None
