+++ pkg/cli/clisqlexec/BUILD.bazel
@@ -19,7 +19,6 @@ go_library(
         "//pkg/cli/clisqlclient",
         "//pkg/sql/lex",
         "//pkg/sql/lexbase",
-        "//pkg/sql/sem/tree",
         "//pkg/sql/sessiondatapb",
         "//pkg/util",
         "//pkg/util/encoding/csv",
@@ -49,7 +48,7 @@ go_test(
         "//pkg/cli/clisqlclient",
         "//pkg/security",
         "//pkg/server",
-        "//pkg/sql/lex",
+        "//pkg/sql/lexbase",
         "//pkg/sql/sem/tree",
         "//pkg/testutils/serverutils",
         "//pkg/testutils/sqlutils",
