+++ pkg/cli/clisqlshell/BUILD.bazel
@@ -19,7 +19,7 @@ go_library(
         "//pkg/cli/democluster/api",
         "//pkg/docs",
         "//pkg/sql",
-        "//pkg/sql/lex",
+        "//pkg/sql/lexbase",
         "//pkg/sql/parser",
         "//pkg/sql/pgwire/pgcode",
         "//pkg/sql/pgwire/pgerror",
