+++ pkg/sql/BUILD.bazel
@@ -304,7 +304,7 @@ go_library(
         "//pkg/sql/gcjob/gcjobnotifier",
         "//pkg/sql/idxusage",
         "//pkg/sql/inverted",
-        "//pkg/sql/lex",
+        "//pkg/sql/lexbase",
         "//pkg/sql/memsize",
         "//pkg/sql/mutations",
         "//pkg/sql/opt",
@@ -557,7 +557,6 @@ go_test(
         "//pkg/sql/execstats",
         "//pkg/sql/flowinfra",
         "//pkg/sql/gcjob",
-        "//pkg/sql/lex",
         "//pkg/sql/lexbase",
         "//pkg/sql/mutations",
         "//pkg/sql/opt/exec/explain",
