+++ pkg/sql/delegate/BUILD.bazel
@@ -42,7 +42,7 @@ go_library(
         "//pkg/settings",
         "//pkg/sql/catalog/catconstants",
         "//pkg/sql/catalog/colinfo",
-        "//pkg/sql/lex",
+        "//pkg/sql/lexbase",
         "//pkg/sql/opt/cat",
         "//pkg/sql/parser",
         "//pkg/sql/pgwire/pgcode",
