+++ pkg/sql/delegate/show_regions.go
@@ -13,7 +13,7 @@ package delegate
 import (
 	"fmt"
 
-	"github.com/cockroachdb/cockroach/pkg/sql/lex"
+	"github.com/cockroachdb/cockroach/pkg/sql/lexbase"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/sqltelemetry"
 	"github.com/cockroachdb/errors"
@@ -70,7 +70,7 @@ FROM [
 LEFT JOIN zones_table ON (r.region = zones_table.region)
 ORDER BY "primary" DESC, "region"`,
 			zonesClause,
-			lex.EscapeSQLString(dbName),
+			lexbase.EscapeSQLString(dbName),
 		)
 		return parse(query)
