+++ pkg/sql/delegate/show_role_grants.go
@@ -15,7 +15,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/cockroachdb/cockroach/pkg/sql/lex"
+	"github.com/cockroachdb/cockroach/pkg/sql/lexbase"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 )
 
@@ -35,7 +35,7 @@ SELECT role AS role_name,
 	if n.Roles != nil {
 		var roles []string
 		for _, r := range n.Roles.ToStrings() {
-			roles = append(roles, lex.EscapeSQLString(r))
+			roles = append(roles, lexbase.EscapeSQLString(r))
 		}
 		fmt.Fprintf(&query, ` WHERE "role" IN (%s)`, strings.Join(roles, ","))
 	}
@@ -51,7 +51,7 @@ SELECT role AS role_name,
 
 		var grantees []string
 		for _, g := range n.Grantees.ToStrings() {
-			grantees = append(grantees, lex.EscapeSQLString(g))
+			grantees = append(grantees, lexbase.EscapeSQLString(g))
 		}
 		fmt.Fprintf(&query, ` member IN (%s)`, strings.Join(grantees, ","))
