+++ pkg/sql/lexbase/BUILD.bazel
@@ -13,14 +13,20 @@ go_library(
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/sql/lexbase",
     visibility = ["//visibility:public"],
-    deps = ["@org_golang_x_text//unicode/norm"],
+    deps = [
+        "//pkg/util/stringencoding",
+        "@org_golang_x_text//unicode/norm",
+    ],
 )
 
 go_test(
     name = "lexbase_test",
     size = "small",
     srcs = ["encode_test.go"],
-    deps = [":lexbase"],
+    deps = [
+        ":lexbase",
+        "//pkg/sql/parser",
+    ],
 )
 
 # Define the target to auto-generate our list of keywords from the grammar file.
