+++ pkg/sql/sem/tree/comment_on_column.go
@@ -10,7 +10,7 @@
 
 package tree
 
-import "github.com/cockroachdb/cockroach/pkg/sql/lex"
+import "github.com/cockroachdb/cockroach/pkg/sql/lexbase"
 
 // CommentOnColumn represents an COMMENT ON COLUMN statement.
 type CommentOnColumn struct {
@@ -29,7 +29,7 @@ func (n *CommentOnColumn) Format(ctx *FmtCtx) {
 		if ctx.flags.HasFlags(FmtHideConstants) {
 			ctx.WriteByte('_')
 		} else {
-			lex.EncodeSQLStringWithFlags(&ctx.Buffer, *n.Comment, ctx.flags.EncodeFlags())
+			lexbase.EncodeSQLStringWithFlags(&ctx.Buffer, *n.Comment, ctx.flags.EncodeFlags())
 		}
 	} else {
 		ctx.WriteString("NULL")
