+++ pkg/testutils/sqlutils/zone.go
@@ -16,7 +16,7 @@ import (
 	"testing"
 
 	"github.com/cockroachdb/cockroach/pkg/config/zonepb"
-	"github.com/cockroachdb/cockroach/pkg/sql/lex"
+	"github.com/cockroachdb/cockroach/pkg/sql/lexbase"
 	"github.com/cockroachdb/cockroach/pkg/util/protoutil"
 )
 
@@ -60,7 +60,7 @@ func DeleteZoneConfig(t testing.TB, sqlDB *SQLRunner, target string) {
 func SetZoneConfig(t testing.TB, sqlDB *SQLRunner, target string, config string) {
 	t.Helper()
 	sqlDB.Exec(t, fmt.Sprintf("ALTER %s CONFIGURE ZONE = %s",
-		target, lex.EscapeSQLString(config)))
+		target, lexbase.EscapeSQLString(config)))
 }
 
 // TxnSetZoneConfig updates the specified zone config through the SQL interface
@@ -68,7 +68,7 @@ func SetZoneConfig(t testing.TB, sqlDB *SQLRunner, target string, config string)
 func TxnSetZoneConfig(t testing.TB, sqlDB *SQLRunner, txn *gosql.Tx, target string, config string) {
 	t.Helper()
 	_, err := txn.Exec(fmt.Sprintf("ALTER %s CONFIGURE ZONE = %s",
-		target, lex.EscapeSQLString(config)))
+		target, lexbase.EscapeSQLString(config)))
 	if err != nil {
 		t.Fatal(err)
 	}
